requirements.txt
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
pyyaml>=6.0
tqdm>=4.66.0
optuna>=3.6.0
mne>=1.6.0
nibabel>=5.2.0
nilearn>=0.10.0
monai>=1.3.0
einops>=0.7.0
mamba-ssm>=2.2.4
timm>=0.9.16
matplotlib>=3.8.0
scikit-learn>=1.4.0
